requirements.txt
Flask
stripe
gunicorn
psycopg2-binary